ma: Move to package io.prometheus.client
